requirements.txt
pandas>=2.0
numpy>=1.24
lightgbm>=4.0
scikit-learn>=1.3
optuna>=3.4
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
pyyaml>=6.0
tqdm>=4.66
joblib>=1.3
matplotlib>=3.8
seaborn>=0.13
pytest>=7.4
